YCC #7.4
Protecting Video Routes

+++ controllers/videoController.js
@@ -52,7 +52,6 @@ export const videoDetail = async (req, res) => {
     } = req;
     try {
         const video = await Video.findById(id).populate("creator");
-        console.log(video);
         res.render("videoDetail", { pageTitle : video.title, video });
     } catch (error) {
         res.redirect(routes.home);
@@ -66,7 +65,11 @@ export const getEditVideo = async (req, res) => {
     } = req;
     try {
         const video = await Video.findById(id);
-        res.render("editVideo", { pageTitle : `Edit ${video.title}`, video });
+        if(video.creator !== req.user.id){
+            throw Error();
+        } else {
+            res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+        }
     } catch (error) {
         res.redirect(routes.home);
     }
@@ -92,11 +95,15 @@ export const deleteVideo = async (req, res) => {
         params: {id}
     } = req;
     try {
-        await Video.findByIdAndRemove({ _id: id});
-        res.redirect(routes.home);
+        const video = await Video.findById(id);
+        if(video.creator !== req.user.id){
+            throw Error();
+        } else {
+            await Video.findByIdAndRemove({ _id: id});
+        }
     } catch (error) {
         console.log(error);
-        res.redirect(routes.home);
     }
-}
+    res.redirect(routes.home);
+};
